+++ Packages/idv.jlchntoz.vvmw/Editor/VVMW/SimpleEditor.cs
@@ -1,5 +1,4 @@
 ﻿using UnityEditor;
-using UdonSharpEditor;
 
 namespace JLChnToZ.VRC.VVMW.Editors {
     [CustomEditor(typeof(VizVidBehaviour), true)]