+++ CHANGELOG.md
@@ -8,10 +8,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - Render Texture as direct video screen targets. The core will blit the screen to these RTs every frame when the video is playing.
 - Fix aspect ratio button in video screen configurator.
+- Global settings window.
 
 ### Changed
 - Canvas Utility now able to handle multiple hierarchy of non-uniform scale adjustment, making sure the UI don't stretched.
 - Updated exhibition setup to use new active region auto play component.
+- Adding active region will now spawns active region manager if none on scene.
+
+### Fixed
+- Standalone queue list / history list button not collapsing playlist selection.
 
 ## 1.7.0-beta.1 - 2026-04-13
 ### Added